pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Rosehack UCR <rosehackucr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
pytz = "^2023.3"
google-auth = "^2.21.0"
requests = "^2.31.0"
google-auth-oauthlib = "^1.0.0"
google-api-python-client = "^2.15.0"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
yapf = "^0.40.1"

[tool.poetry.scripts]
bot = "bot.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
